rename ApeClient to ts-rest-adapter
