+++ frontend/src/ts/ape/adapters/ts-rest-adapter.ts
